Feat: Add auth to routes patch and delete

+++ src/companys/repositories/companys.repository.ts
@@ -1,12 +1,17 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { CreateCompanyDto } from '../dto/create-company.dto';
 import { UpdateCompanyDto } from '../dto/update-company.dto';
 import { CompanyEntity } from '../entities/company.entity';
+import { JwtService } from '@nestjs/jwt';
+import { Request } from 'express';
 
 @Injectable()
 export class CompanysRepository {
-  constructor(private readonly prisma: PrismaService) {}
+  constructor(
+    private readonly prisma: PrismaService,
+    private readonly jwtService: JwtService,
+  ) {}
   async create(createCompanyDto: CreateCompanyDto): Promise<CompanyEntity> {
     return this.prisma.company.create({
       data: createCompanyDto,
@@ -177,21 +182,81 @@ export class CompanysRepository {
       },
     });
   }
+  async resume(page: number) {
+    const itemsPerPage = 10;
+    const skip = (page - 1) * itemsPerPage;
 
-  async update(id: number, updateCompanyDto: UpdateCompanyDto) {
-    return this.prisma.company.update({
-      where: {
-        id,
+    return this.prisma.company.findMany({
+      skip,
+      take: itemsPerPage,
+      select: {
+        id: true,
+        name: true,
+        ceo: true,
+        headquarters: true,
+        founding_date: true,
+        more_info: true,
+        picsId: true,
+        pictures: {
+          select: {
+            id: true,
+            name: true,
+            banners: {
+              select: {
+                id: true,
+                url: true,
+              },
+            },
+            profiles: {
+              select: {
+                id: true,
+                url: true,
+              },
+            },
+          },
+        },
       },
-      data: updateCompanyDto,
     });
   }
 
-  async remove(id: number) {
-    return this.prisma.company.delete({
-      where: {
-        id,
-      },
-    });
+  async update(id: number, updateCompanyDto: UpdateCompanyDto, req: Request) {
+    try {
+      const cookie = req.cookies['jwt'];
+      const data = await this.jwtService.verifyAsync(cookie);
+
+      if (!data) {
+        throw new BadRequestException('Unauthorized');
+      }
+      //Posteriormente adicionar Role e verificar se a role do usuário é admin ou moderador
+
+      return this.prisma.company.update({
+        where: {
+          id,
+        },
+        data: updateCompanyDto,
+      });
+    } catch (e) {
+      throw new BadRequestException('Unauthenticated');
+    }
+  }
+
+  async remove(id: number, req: Request) {
+    try {
+      const cookie = req.cookies['jwt'];
+      const data = await this.jwtService.verifyAsync(cookie);
+
+      if (!data) {
+        throw new BadRequestException('Unauthorized');
+      }
+      //Posteriormente adicionar Role e verificar se a role do usuário é admin ou moderador
+
+      return this.prisma.company.delete({
+        where: {
+          id,
+        },
+      });
+    } catch (e) {
+      throw new BadRequestException('Unauthenticated');
+    }
   }
 }
